create and move liquidate_position endpoint
